movie_service/etl/requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
requests==2.31.0